Move non-shared components closer to point of use
