+++ app/page.tsx
@@ -1,6 +1,35 @@
-import { stripBlogPostBody } from "lib/post";
+import { PostMetadata } from "lib/post";
 import { siteData } from "lib/site-data";
-import PostListing from "components/post-listing";
+import PostCard from "components/post-card";
+import PressReleaseListing from "components/press-releases";
+
+const Home = () => {
+  const authors = siteData.authors;
+  const [firstPost, ...otherPosts] = siteData.posts;
+  const authorOf = (post: PostMetadata) =>
+    authors.find((a) => a.id === post.authorId)!;
+  return (
+    <div className="post-listing-row">
+      <div className="main-post">
+        <PostCard
+          post={firstPost}
+          author={authorOf(firstPost)}
+          showCover={true}
+        />
+      </div>
+
+      <div className="press-release-box">
+        <PressReleaseListing />
+      </div>
+
+      {otherPosts.map((post) => (
+        <div className="post-listing-post" key={post.path}>
+          <PostCard post={post} author={authorOf(post)} />
+        </div>
+      ))}
+    </div>
+  );
+};
 
 export const metadata = {
   title: "Blog Česko.Digital",
@@ -15,16 +44,6 @@ export const metadata = {
       },
     ],
   },
-};
-
-const Home = () => {
-  const { authors } = siteData;
-  const posts = siteData.posts.map(stripBlogPostBody);
-  return (
-    <div>
-      <PostListing posts={posts} authors={authors} />
-    </div>
-  );
 };
 
 export default Home;
